orchestrator/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
GitPython==3.1.43
